feat(client): refactor perps session actions
